+++ pkg/metrics/instance/configstore/kv/consul/client.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/go-kit/kit/log/level"
+	"github.com/grafana/agent/pkg/metrics/instance/configstore/kv/pair"
 	consul "github.com/hashicorp/consul/api"
 	"github.com/hashicorp/go-cleanhttp"
 	"github.com/weaveworks/common/instrument"
@@ -264,10 +265,12 @@ func (c *Client) WatchKey(ctx context.Context, key string, f func(interface{}) b
 // Values in Consul are assumed to be JSON. This function blocks until the context is cancelled.
 func (c *Client) WatchPrefix(ctx context.Context, prefix string, f func(string, interface{}) bool) {
 	var (
-		backoff = util.NewBackoff(ctx, backoffConfig)
-		index   = uint64(0)
-		limiter = c.createRateLimiter()
+		backoff  = util.NewBackoff(ctx, backoffConfig)
+		index    = uint64(0)
+		limiter  = c.createRateLimiter()
+		lastKeys = map[string]struct{}{}
 	)
+
 	for backoff.Ongoing() {
 		err := limiter.Wait(ctx)
 		if err != nil {
@@ -287,6 +290,7 @@ func (c *Client) WatchPrefix(ctx context.Context, prefix string, f func(string, 
 		}
 
 		kvps, meta, err := c.kv.List(prefix, queryOptions.WithContext(ctx))
+
 		// kvps being nil here is not an error -- quite the opposite. Consul returns index,
 		// which makes next query blocking, so there is no need to detect this and act on it.
 		if err != nil {
@@ -301,7 +305,13 @@ func (c *Client) WatchPrefix(ctx context.Context, prefix string, f func(string, 
 			continue
 		}
 
+		newKeys := make(map[string]struct{}, len(kvps))
+
 		for _, kvp := range kvps {
+			// Keep track of all returned keys. This *must* be done before the check
+			// below otherwise an unchanged key will be treated as deleted.
+			newKeys[kvp.Key] = struct{}{}
+
 			// We asked for values newer than 'index', but Consul returns all values below given prefix,
 			// even those that haven't changed. We don't need to report all of them as updated.
 			if index > 0 && kvp.ModifyIndex <= index && kvp.CreateIndex <= index {
@@ -318,12 +328,24 @@ func (c *Client) WatchPrefix(ctx context.Context, prefix string, f func(string, 
 			}
 		}
 
+		// Send deleted events using keys that have disappeared betwwn the last
+		// List and this List.
+		for key := range lastKeys {
+			_, exist := newKeys[key]
+			if !exist {
+				if !f(key, nil) {
+					return
+				}
+			}
+		}
+
 		index = newIndex
+		lastKeys = newKeys
 	}
 }
 
 // List implements kv.List.
-func (c *Client) List(ctx context.Context, prefix string) ([]string, error) {
+func (c *Client) List(ctx context.Context, prefix string) ([]pair.KVP, error) {
 	options := &consul.QueryOptions{
 		AllowStale:        !c.cfg.ConsistentReads,
 		RequireConsistent: c.cfg.ConsistentReads,
@@ -333,11 +355,20 @@ func (c *Client) List(ctx context.Context, prefix string) ([]string, error) {
 		return nil, err
 	}
 
-	keys := make([]string, 0, len(pairs))
+	res := make([]pair.KVP, 0, len(pairs))
 	for _, kvp := range pairs {
-		keys = append(keys, kvp.Key)
+		value, err := c.codec.Decode(kvp.Value)
+		if err != nil {
+			level.Error(util_log.Logger).Log("msg", "error decoding list of values for prefix:key", "prefix", prefix, "key", kvp.Key, "err", err)
+			continue
+		}
+
+		res = append(res, pair.KVP{
+			Key:   kvp.Key,
+			Value: value,
+		})
 	}
-	return keys, nil
+	return res, nil
 }
 
 // Get implements kv.Get.