+++ pkg/metrics/instance/configstore/remote.go
@@ -2,17 +2,9 @@ package configstore
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"net/http"
 	"strings"
 	"sync"
-
-	"github.com/weaveworks/common/instrument"
-
-	"github.com/hashicorp/go-cleanhttp"
-
-	"github.com/hashicorp/consul/api"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
@@ -20,20 +12,7 @@ import (
 	"github.com/grafana/agent/pkg/metrics/instance/configstore/kv"
 	"github.com/grafana/agent/pkg/util"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
 )
-
-/***********************************************************************************************************************
-The consul code skipping the cortex handler is due to performance issue with a large number of configs and overloading
-consul. See issue https://github.com/grafana/agent/issues/789. The long term method will be to refactor and extract
-the cortex code so other stores can also benefit from this. @mattdurham
-***********************************************************************************************************************/
-
-var consulRequestDuration = instrument.NewHistogramCollector(promauto.NewHistogramVec(prometheus.HistogramOpts{
-	Name:    "agent_configstore_consul_request_duration_seconds",
-	Help:    "Time spent on consul requests when listing configs.",
-	Buckets: prometheus.DefBuckets,
-}, []string{"operation", "status_code"}))
 
 // Remote loads instance files from a remote KV store. The KV store
 // can be swapped out in real time.
@@ -42,7 +21,7 @@ type Remote struct {
 	reg *util.Unregisterer
 
 	kvMut    sync.RWMutex
-	kv       *agentRemoteClient
+	kv       kv.Client
 	reloadKV chan struct{}
 
 	cancelCtx  context.Context
@@ -50,14 +29,6 @@ type Remote struct {
 
 	configsMut sync.Mutex
 	configsCh  chan WatchEvent
-}
-
-// agentRemoteClient is a simple wrapper to allow the shortcircuit of consul, while being backwards compatible with non
-// consul kv stores
-type agentRemoteClient struct {
-	kv.Client
-	consul *api.Client
-	config kv.Config
 }
 
 // NewRemote creates a new Remote store that uses a Key-Value client to store
@@ -99,49 +70,23 @@ func (r *Remote) ApplyConfig(cfg kv.Config, enable bool) error {
 	r.reg.UnregisterAll()
 
 	if !enable {
-		r.setClient(nil, nil, kv.Config{})
+		r.setClient(nil)
 		return nil
 	}
 
 	cli, err := kv.NewClient(cfg, GetCodec(), kv.RegistererWithKVName(r.reg, "agent_configs"))
-	// This is a hack to get a consul client, the client above has it embedded but its not exposed
-	var consulClient *api.Client
-	if cfg.Store == "consul" {
-		consulClient, err = api.NewClient(&api.Config{
-			Address: cfg.Consul.Host,
-			Token:   cfg.Consul.ACLToken,
-			Scheme:  "http",
-			HttpClient: &http.Client{
-				Transport: cleanhttp.DefaultPooledTransport(),
-				// See https://blog.cloudflare.com/the-complete-guide-to-golang-net-http-timeouts/
-				Timeout: cfg.Consul.HTTPClientTimeout,
-			},
-		})
-		if err != nil {
-			return err
-		}
-
-	}
 	if err != nil {
 		return fmt.Errorf("failed to create kv client: %w", err)
 	}
 
-	r.setClient(cli, consulClient, cfg)
+	r.setClient(cli)
 	return nil
 }
 
 // setClient sets the active client and notifies run to restart the
 // kv watcher.
-func (r *Remote) setClient(client kv.Client, consulClient *api.Client, config kv.Config) {
-	if client == nil && consulClient == nil {
-		r.kv = nil
-	} else {
-		r.kv = &agentRemoteClient{
-			Client: client,
-			consul: consulClient,
-			config: config,
-		}
-	}
+func (r *Remote) setClient(client kv.Client) {
+	r.kv = client
 	r.reloadKV <- struct{}{}
 }
 
@@ -174,7 +119,7 @@ Outer:
 	}
 }
 
-func (r *Remote) watchKV(ctx context.Context, client *agentRemoteClient) {
+func (r *Remote) watchKV(ctx context.Context, client kv.Client) {
 	// Edge case: client was unset, nothing to do here.
 	if client == nil {
 		level.Info(r.log).Log("msg", "not watching the KV, none set")
@@ -214,39 +159,15 @@ func (r *Remote) List(ctx context.Context) ([]string, error) {
 		return nil, ErrNotConnected
 	}
 
-	return r.kv.List(ctx, "")
-}
-
-// listConsul returns Key Value Pairs instead of []string
-func (r *Remote) listConsul(ctx context.Context) (api.KVPairs, error) {
-	if r.kv == nil {
-		return nil, ErrNotConnected
-	}
-
-	var pairs api.KVPairs
-	options := &api.QueryOptions{
-		AllowStale:        !r.kv.config.Consul.ConsistentReads,
-		RequireConsistent: r.kv.config.Consul.ConsistentReads,
-	}
-	// This is copied from cortex list so that stats stay the same
-	err := instrument.CollectedRequest(ctx, "List", consulRequestDuration, instrument.ErrorCode, func(ctx context.Context) error {
-		var err error
-		pairs, _, err = r.kv.consul.KV().List(r.kv.config.Prefix, options.WithContext(ctx))
-		return err
-	})
-
+	kvps, err := r.kv.List(ctx, "")
 	if err != nil {
 		return nil, err
 	}
-	// This mirrors the previous behavior of returning a blank array as opposed to nil.
-	if pairs == nil {
-		blankPairs := make(api.KVPairs, 0)
-		return blankPairs, nil
+	keys := make([]string, len(kvps))
+	for i, kvp := range kvps {
+		keys[i] = kvp.Key
 	}
-	for _, kvp := range pairs {
-		kvp.Key = strings.TrimPrefix(kvp.Key, r.kv.config.Prefix)
-	}
-	return pairs, nil
+	return keys, nil
 }
 
 // Get retrieves an individual config from the KV store.
@@ -347,110 +268,31 @@ func (r *Remote) all(ctx context.Context, keep func(key string) bool) (<-chan in
 		return nil, ErrNotConnected
 	}
 
-	// If we are using a consul client then do the short circuit way, this is done so that we receive all the key value pairs
-	//	in one call then, operate on them in memory. Previously we retrieved the list (which stripped the values)
-	//	then ran a goroutine to get each individual value from consul. In situations with an extremely large number of
-	//	configs this overloaded the consul instances. This reduces that to one call, that was being made anyways.
-	if r.kv.consul != nil {
-		return r.allConsul(ctx, keep)
-	}
-	return r.allOther(ctx, keep)
-
-}
-
-// allConsul is ONLY usable when consul is the keystore. This is a performance improvement in using the client directly
-//	instead of the cortex multi store kv interface. That interface returns the list then each value must be retrieved
-//	individually. This returns all the keys and values in one call and works on them in memory
-func (r *Remote) allConsul(ctx context.Context, keep func(key string) bool) (<-chan instance.Config, error) {
-	if r.kv.consul == nil {
-		level.Error(r.log).Log("err", "allConsul called but consul client nil")
-		return nil, errors.New("allConsul called but consul client nil")
-	}
-	var configs []*instance.Config
-	c := GetCodec()
-
-	pairs, err := r.listConsul(ctx)
-
-	if err != nil {
-		return nil, err
-	}
-	for _, kvp := range pairs {
-		if keep != nil && !keep(kvp.Key) {
-			level.Debug(r.log).Log("msg", "skipping key that was filtered out", "key", kvp.Key)
-			continue
-		}
-		value, err := c.Decode(kvp.Value)
-		if err != nil {
-			level.Error(r.log).Log("msg", "failed to decode config from store", "key", kvp.Key, "err", err)
-			continue
-		}
-		if value == nil {
-			// Config was deleted since we called list, skip it.
-			level.Debug(r.log).Log("msg", "skipping key that was deleted after list was called", "key", kvp.Key)
-			continue
-		}
-
-		cfg, err := instance.UnmarshalConfig(strings.NewReader(value.(string)))
-		if err != nil {
-			level.Error(r.log).Log("msg", "failed to unmarshal config from store", "key", kvp.Key, "err", err)
-			continue
-		}
-		configs = append(configs, cfg)
-	}
-	ch := make(chan instance.Config, len(configs))
-	for _, cfg := range configs {
-		ch <- *cfg
-	}
-	close(ch)
-	return ch, nil
-}
-
-func (r *Remote) allOther(ctx context.Context, keep func(key string) bool) (<-chan instance.Config, error) {
-	if r.kv == nil {
-		return nil, ErrNotConnected
-	}
-
-	keys, err := r.kv.List(ctx, "")
+	kvps, err := r.kv.List(ctx, "")
 	if err != nil {
 		return nil, fmt.Errorf("failed to list configs: %w", err)
 	}
 
-	ch := make(chan instance.Config)
+	// NOTE(rfratto): ch must be buffered for everything we're going to write to
+	// it, otherwise this goroutine will deadlock.
+	//
+	// Data written to the buffered channel will still be available for reading
+	// even after the channel is closed.
+	ch := make(chan instance.Config, len(kvps))
+	defer close(ch)
 
-	var wg sync.WaitGroup
-	wg.Add(len(keys))
-	go func() {
-		wg.Wait()
-		close(ch)
-	}()
+	for _, kvp := range kvps {
+		if keep != nil && !keep(kvp.Key) {
+			level.Debug(r.log).Log("msg", "skipping key that was filtered out", "key", kvp.Key)
+			continue
+		}
 
-	for _, key := range keys {
-		go func(key string) {
-			defer wg.Done()
-
-			if keep != nil && !keep(key) {
-				level.Debug(r.log).Log("msg", "skipping key that was filtered out", "key", key)
-				return
-			}
-
-			// TODO(rfratto): retries might be useful here
-			v, err := r.kv.Get(ctx, key)
-			if err != nil {
-				level.Error(r.log).Log("msg", "failed to get config with key", "key", key, "err", err)
-				return
-			} else if v == nil {
-				// Config was deleted since we called list, skip it.
-				level.Debug(r.log).Log("msg", "skipping key that was deleted after list was called", "key", key)
-				return
-			}
-
-			cfg, err := instance.UnmarshalConfig(strings.NewReader(v.(string)))
-			if err != nil {
-				level.Error(r.log).Log("msg", "failed to unmarshal config from store", "key", key, "err", err)
-				return
-			}
-			ch <- *cfg
-		}(key)
+		cfg, err := instance.UnmarshalConfig(strings.NewReader(kvp.Value.(string)))
+		if err != nil {
+			level.Error(r.log).Log("msg", "failed to unmarshal config from store", "key", kvp.Key, "err", err)
+			continue
+		}
+		ch <- *cfg
 	}
 
 	return ch, nil